python/myscripts_python/requirements.txt
ansible==2.9.9
asyncio==3.4.3
asyncssh==1.18.0
click==7.1.2
Jinja2==2.11.2
napalm==3.0.1
netdev==0.9.3
netmiko==3.1.1
ntc-templates==1.4.2
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
textfsm==1.1.0
junos-eznc==2.4.1
xmltodict==0.12.0

